pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "byos"
version = "0.0.0"
description = "BYOS (Bring Your Own State): research prototype for stateful long-context LLMs"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
dependencies = [
  "torch",
  "numpy",
  "pyyaml",
  "tqdm",
  "datasets<4.0.0",
  "transformers",
  "tokenizers",
  "sentencepiece",
]

[project.optional-dependencies]
comet = ["comet-ml"]

[tool.setuptools.packages.find]
where = ["."]
include = ["BYOS*"]
